+++ NEWS.md
@@ -1,3 +1,5 @@
+# dtplyr (development version)
+
 # dtplyr 1.3.1
 
 * Fix for failing R CMD check.