+++ NEWS.md
@@ -1,5 +1,10 @@
 # dtplyr (development version)
 
+## New features
+
+* `consecutive_id()` is now mapped to `data.table::rleid()`.
+  Note: `rleid()` only accepts vector inputs and cannot be used with data frame inputs.
+
 # dtplyr 1.3.1
 
 * Fix for failing R CMD check.