+++ NEWS.md
@@ -19,6 +19,8 @@
 
 * `.by` no longer alters grouping in prior steps (#439)
 
+* Arguments to `$` calls are no longer prepended with `..`  (#434)
+
 # dtplyr 1.3.1
 
 * Fix for failing R CMD check.