+++ NEWS.md
@@ -19,7 +19,7 @@
 
 * `.by` no longer alters grouping in prior steps (#439)
 
-* Arguments to `$` calls are no longer prepended with `..`  (#434)
+* Arguments to `$` and `[[` calls are no longer prepended with `..`  (#434)
 
 # dtplyr 1.3.1