+++ NEWS.md
@@ -1,4 +1,4 @@
-# dtplyr (development version)
+# dtplyr 1.3.2
 
 * `R CMD check` fixes